+++ webapp/components/play-form.tsx
@@ -54,8 +54,8 @@ export function PlayForm() {
             setError("サインインしてください。");
         }
         if (selectedContent && user) {
-            if (isLimited && !joinWord.trim()) {
-                setError("限定部屋では入室の言葉を入力してください。");
+            if (isLimited && !joinWord) {
+                setError("限定部屋を作成する場合、入室の言葉が必要です。");
                 return;
             }
             setIsSending(true);
@@ -68,9 +68,13 @@ export function PlayForm() {
                 joinWord,
             });
             if (res.ok) {
-                redirect(
-                    `/play/${res.playId}?${messageKey}=${messages.play.registerSuccessful}`,
-                );
+                const query = new URLSearchParams({
+                    messageKey: messages.play.registerSuccessful,
+                });
+                if (isLimited) {
+                    query.set("inviteHash", res.inviteHash!);
+                }
+                redirect(`/play/${res.playId}?${query.toString()}`);
             } else {
                 switch (res.reason) {
                     case "InvalidParams":
@@ -175,7 +179,7 @@ export function PlayForm() {
                                 <FormControlLabel
                                     value="limited"
                                     control={<Radio />}
-                                    label="限定: 部屋一覧には表示されますが、一覧から入るには入室の言葉が必要です。"
+                                    label="限定: 部屋一覧には表示されますが、入室の言葉がないと入れません。"
                                 />
                             </RadioGroup>
                         </Box>
@@ -193,7 +197,7 @@ export function PlayForm() {
                                             maxLength: 100,
                                         },
                                     }}
-                                    helperText="限定部屋に一覧から入るときに必要な言葉です。"
+                                    helperText="部屋一覧から入室するときに必要な言葉です。"
                                 />
                             </Box>
                         ) : null}