+++ webapp/components/play-list.tsx
@@ -25,6 +25,7 @@ import {
     Skeleton,
     Stack,
     TextField,
+    Tooltip,
     Typography,
     useTheme,
 } from "@mui/material";
@@ -35,6 +36,7 @@ import {
     Search,
     SportsEsports,
     Link as LinkIcon,
+    Lock,
 } from "@mui/icons-material";
 import { PlayInfo } from "@/lib/types";
 import { usePlayList } from "@/lib/client/usePlayList";
@@ -101,12 +103,25 @@ function PlayGrid({ list }: { list: PlayInfo[] }) {
                                 {info.playName}
                             </Typography>
                             {info.isLimited ? (
-                                <Chip
-                                    size="small"
-                                    label="限定"
-                                    color="warning"
-                                    sx={{ width: "fit-content", mb: 1 }}
-                                />
+                                <Tooltip
+                                    arrow
+                                    title="この部屋に入室するには「入室の言葉」が必要です。"
+                                >
+                                    <Stack spacing={0.5} direction="row">
+                                        <Lock fontSize="small" />
+                                        <Typography
+                                            variant="body2"
+                                            color={theme.palette.text.secondary}
+                                            sx={{
+                                                px: 1,
+                                                py: 0.5,
+                                                borderRadius: 1,
+                                            }}
+                                        >
+                                            限定
+                                        </Typography>
+                                    </Stack>
+                                </Tooltip>
                             ) : null}
                             <Typography
                                 variant="body2"